Better handle rendering of small maps.
Fixes #3. Swaps order of evaluation of bounds in update_camera(), which isn't strictly necessary and forces the map to the upper left corner of the panel; not swapping them would force the map to the lower right, and it'd take a bit of math that *only* fires for small maps to center.